Final Touches
Added the rest of the cases for the GUI's switches that made us of the tax calculator as well as methods for the GUI to use, mainly for the admin to use to see user's info or to remove users. Added a method I forgot to add to the tax calculator to allow the users to see their info and taxes owed and commented all the code that I made to better explain how it worked.