Update and create
Did some changes in users because I decided to create a new class called modifyProfile to update the data from the database. Also changed and commented GUI and database. For Gui I implemented the new method that I created to modify the profile and commenting the code. And in the database I created the new method for the modify profiles and also commented the code.